+++ GPy/plotting/matplot_dep/base_plots.py
@@ -103,7 +103,7 @@ def gradient_fill(x, percentiles, ax=None, fignum=None, **kwargs):
                 continue
 
             N = len(xslice)
-            X = np.zeros((2 * N + 2, 2), np.float)
+            X = np.zeros((2 * N + 2, 2), float)
 
             # the purpose of the next two lines is for when y2 is a
             # scalar like 0 and we want the fill to go all the way
